move reflection app tests
